requirements.txt
absl-py==0.11.0
asgiref==3.3.1
astunparse==1.6.3
cachetools==4.2.0
certifi==2020.12.5
chardet==4.0.0
Django==3.1.4
flatbuffers==1.12
gast==0.3.3
google-auth==1.24.0
google-auth-oauthlib==0.4.2
google-pasta==0.2.0
grpcio==1.32.0
h5py==2.10.0
idna==2.10
joblib==1.0.0
Keras==2.3.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
livereload==2.6.3
Markdown==3.3.3
numpy==1.19.3
oauthlib==3.1.0
opt-einsum==3.3.0
pandas==1.1.5
protobuf==3.14.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
python-dateutil==2.8.1
pytz==2020.4
PyYAML==5.3.1
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.6
scikit-learn==0.23.1
scipy==1.5.4
six==1.15.0
sqlparse==0.4.1
tensorboard==2.4.0
tensorboard-plugin-wit==1.7.0

tensorflow-estimator==2.4.0
termcolor==1.1.0
threadpoolctl==2.1.0
tornado==6.1
typing-extensions==3.7.4.3
urllib3==1.26.2
Werkzeug==1.0.1
wrapt==1.12.1
